+++ src/helpers/generateUUID.js
@@ -1,0 +1,4 @@
+export default () => {
+  const uuid = crypto.randomUUID();
+  return uuid.substring(0, uuid.indexOf('-'));
+};